[IMP] WIP: ffmpeg formatting
